[Change] Decouple JDK running Gradle from JDK building project
Gradle will by default use the same JDK for building the project as it
does for running Gradle. Newer JDKs are assumed to be more performant,
contain fewer bugs and to come with desirable features that we would
like to utilize in our build, but doing so should not endanger
backwards compatibility for the plugin itself. We can do that now by
utilizing Gradle toolchains, which allows us to use one JDK/JRE for
running Gradle and the other for building the project, i.e. executing
`javac`.

Doing this requires us to make multiple JDKs available to the Gradle
build action for each individual execution of the build.

Additionally, since `--release=8` is no longer possible from this change, the `java8unsupported` shim has also been removed. `options.release` only effects builds where Java 16+ is in use.