[Change] Added support for `deprecated.json`
Port of the system Mojang Introduced in Minecraft Snapshot 24w33a (The first snapshot for 1.21.2 released last month)

This system works 1:1 with how Mojang parses it, whilst allowing any mod to use it,
regardless of MC version, or if you are using a .lang or .json file (Albeit, `deprecated.json` must be a json file)

The Logic Outline is as follows:
- `deprecated.json` is only read from the internal mod files -- For MC, this means you can't override it with a resource pack
- The JSON contains two fields, a string array field called `removed`, and a `Map<String, String>` for `renamed`
- `removed` simply defines strings you wish to remove from the final translation mappings
- `renamed` defines instances where you wish to transfer a value bound to the `oldKey` to the `newKey` (IE you're renaming translation keys)
- `deprecated.json` is a global file, so the file itself is only loaded once in the logic flow, populating the `removed` and `renamed` fields, for when a language is changed

This file is completely optional, and is primarily made to avoid having to adjust services like Crowdin too often,
or otherwise cause lost translations from now on.

Also note: This feature can be disabled through the chain-method `setCheckDeprecations`,
if it feels unnecessary, or if you have another system in use for your workflow.